Refactor project structure: rename main class to Main (before, APP); move exception classes and properties to proper packages; update FXML utils and controller structure

+++ GenericLists/src/main/java/mycompany/genericlists/models/Student.java
@@ -10,12 +10,12 @@ package mycompany.genericlists.models;
  */
 public class Student {
     private String name;
-    private int id;
+    private long id;
     
     
     /* ------- Constructors ------- */
     
-    public Student(String name, int id) {
+    public Student(String name, long id) {
         this.name = name;
         this.id = id;
     }
@@ -32,11 +32,11 @@ public class Student {
         this.name = name;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
